Delete GUI5_ManualDroneControl directory
This directory, as well as every other directory beginning with "GUI5", appear to only contain unused copies of files in the root directory (and the GUI_Pics directory, in the case of GUI5_Manual DroneControl/images). My guess is that an attempt was made to move the qml files from the root directory into different folders at some point, but other commits reintroduced the files in the root directory. In any case, only the files in the root directory have been kept up to date. I have deleted all directories starting with "GUI5" to prevent future confusions.